Fix several subtle code errors with contexts, and make LocalContext stricter.
Some regexes had major escaping problems and just happened to have
almost-correct behavior. Also, some calls to re.match were not properly
checking to the end of the string.

Contexts now more directly check if data is valid, such as checking if the
selected text matches a filename even if it doesn't match the LocalContext
regex. Finally, brackets, parens, and equals sign terminate the LocalContext
match if they're not escaped.